+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/schema/generator/SchemaGenerator.java
@@ -43,6 +43,11 @@ import io.ballerina.compiler.api.symbols.TypeSymbol;
 import io.ballerina.compiler.api.symbols.UnionTypeSymbol;
 import io.ballerina.compiler.api.symbols.resourcepath.PathSegmentList;
 import io.ballerina.compiler.api.symbols.resourcepath.util.PathSegment;
+import io.ballerina.compiler.syntax.tree.Node;
+import io.ballerina.compiler.syntax.tree.ObjectConstructorExpressionNode;
+import io.ballerina.compiler.syntax.tree.ServiceDeclarationNode;
+import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
 import io.ballerina.stdlib.graphql.compiler.schema.types.DefaultDirective;
 import io.ballerina.stdlib.graphql.compiler.schema.types.Description;
 import io.ballerina.stdlib.graphql.compiler.schema.types.Directive;
@@ -56,15 +61,19 @@ import io.ballerina.stdlib.graphql.compiler.schema.types.TypeKind;
 import io.ballerina.stdlib.graphql.compiler.schema.types.TypeName;
 import io.ballerina.stdlib.graphql.compiler.service.InterfaceFinder;
 
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 import static io.ballerina.stdlib.graphql.compiler.Utils.getAccessor;
 import static io.ballerina.stdlib.graphql.compiler.Utils.getEffectiveType;
 import static io.ballerina.stdlib.graphql.compiler.Utils.getEffectiveTypes;
 import static io.ballerina.stdlib.graphql.compiler.Utils.isContextParameter;
 import static io.ballerina.stdlib.graphql.compiler.Utils.isFileUploadParameter;
+import static io.ballerina.stdlib.graphql.compiler.Utils.isFunctionDefinition;
 import static io.ballerina.stdlib.graphql.compiler.Utils.isRemoteMethod;
 import static io.ballerina.stdlib.graphql.compiler.Utils.isResourceMethod;
 import static io.ballerina.stdlib.graphql.compiler.schema.generator.GeneratorUtils.MAP_KEY_ARGUMENT_DESCRIPTION;
@@ -84,18 +93,21 @@ public class SchemaGenerator {
     private static final String IF_ARG_NAME = "if";
     private static final String REASON_ARG_NAME = "reason";
 
-    private final ServiceDeclarationSymbol serviceDeclarationSymbol;
+    private final Node serviceNode;
     private final InterfaceFinder interfaceFinder;
     private final Schema schema;
+    private final SyntaxNodeAnalysisContext context;
 
-    public SchemaGenerator(ServiceDeclarationSymbol serviceDeclarationSymbol, InterfaceFinder interfaceFinder) {
-        this.serviceDeclarationSymbol = serviceDeclarationSymbol;
+    public SchemaGenerator(SyntaxNodeAnalysisContext context, Node serviceNode, InterfaceFinder interfaceFinder,
+                           String description) {
+        this.context = context;
+        this.serviceNode = serviceNode;
         this.interfaceFinder = interfaceFinder;
-        this.schema = new Schema(getDescription(serviceDeclarationSymbol));
+        this.schema = new Schema(description);
     }
 
     public Schema generate() {
-        findRootTypes(this.serviceDeclarationSymbol);
+        findRootTypes(this.serviceNode);
         findIntrospectionTypes();
         return this.schema;
     }
@@ -106,9 +118,9 @@ public class SchemaGenerator {
         addDefaultDirectives();
     }
 
-    private void findRootTypes(ServiceDeclarationSymbol serviceDeclarationSymbol) {
+    private void findRootTypes(Node serviceNode) {
         Type queryType = addType(TypeName.QUERY);
-        for (MethodSymbol methodSymbol : serviceDeclarationSymbol.methods().values()) {
+        for (MethodSymbol methodSymbol : getMethods(serviceNode)) {
             if (isResourceMethod(methodSymbol)) {
                 ResourceMethodSymbol resourceMethodSymbol = (ResourceMethodSymbol) methodSymbol;
                 String accessor = getAccessor(resourceMethodSymbol);
@@ -130,6 +142,35 @@ public class SchemaGenerator {
         if (this.schema.containsType(TypeName.SUBSCRIPTION.getName())) {
             this.schema.setSubscriptionType(this.schema.getType(TypeName.SUBSCRIPTION.getName()));
         }
+    }
+
+    private Collection<? extends MethodSymbol> getMethods(Node node) {
+        if (node.kind() == SyntaxKind.SERVICE_DECLARATION) {
+            return getMethods((ServiceDeclarationNode) node);
+        }
+
+        if (node.kind() == SyntaxKind.OBJECT_CONSTRUCTOR) {
+            return getMethods((ObjectConstructorExpressionNode) node);
+        }
+
+        return new ArrayList<>();
+    }
+
+    private Collection<? extends MethodSymbol> getMethods(
+            ObjectConstructorExpressionNode objectConstructorExpressionNode) {
+        // noinspection OptionalGetWithoutIsPresent
+        return objectConstructorExpressionNode.members().stream()
+                .filter(member -> isFunctionDefinition(member) && context.semanticModel().symbol(member)
+                        .isPresent()).map(methodNode -> (MethodSymbol) context.semanticModel().symbol(methodNode).get())
+                .collect(Collectors.toList());
+    }
+
+    private Collection<? extends MethodSymbol> getMethods(ServiceDeclarationNode serviceDeclarationNode) {
+        // ServiceDeclarationSymbol already validated. Therefore, no need to check isEmpty().
+        // noinspection OptionalGetWithoutIsPresent
+        ServiceDeclarationSymbol serviceDeclarationSymbol = (ServiceDeclarationSymbol) context.semanticModel()
+                .symbol(serviceDeclarationNode).get();
+        return serviceDeclarationSymbol.methods().values();
     }
 
     private Field getField(ResourceMethodSymbol methodSymbol) {